+++ frontend/src/app/main/data/workspace.cljs
@@ -1002,6 +1002,23 @@
               (rx/of (dwe/start-edition-mode id)
                      (dwdp/start-path-edit id)))))))))
 
+(defn select-parent-layer
+  []
+  (ptk/reify ::select-parent-layer
+    ptk/WatchEvent
+    (watch [_ state _]
+      (let [selected (wsh/lookup-selected state)
+            objects (wsh/lookup-page-objects state)
+            shapes-to-select
+            (->> selected
+                 (reduce
+                   (fn [result shape-id]
+                     (let [parent-id (dm/get-in objects [shape-id :parent-id])]
+                       (if (and (some? parent-id)  (not= parent-id uuid/zero))
+                         (conj result parent-id)
+                         (conj result shape-id))))
+                   (d/ordered-set)))]
+        (rx/of (dws/select-shapes shapes-to-select))))))
 
 ;; --- Change Page Order (D&D Ordering)
 
@@ -1114,7 +1131,7 @@
 
 (defn toggle-proportion-lock
   []
-  (ptk/reify ::toggle-propotion-lock
+  (ptk/reify ::toggle-proportion-lock
     ptk/WatchEvent
     (watch [_ state _]
       (let [page-id       (:current-page-id state)
@@ -1371,7 +1388,7 @@
 
 (defn go-to-viewer
   ([] (go-to-viewer {}))
-  ([{:keys [file-id page-id section]}]
+  ([{:keys [file-id page-id section frame-id]}]
    (ptk/reify ::go-to-viewer
      ptk/WatchEvent
      (watch [_ state _]
@@ -1379,7 +1396,9 @@
              pparams {:file-id (or file-id current-file-id)}
              qparams (cond-> {:page-id (or page-id current-page-id)}
                        (some? section)
-                       (assoc :section section))]
+                       (assoc :section section)
+                       (some? frame-id)
+                       (assoc :frame-id frame-id))]
          (rx/of ::dwp/force-persist
                 (rt/nav-new-window* {:rname :viewer
                                      :path-params pparams