+++ frontend/src/app/main/data/workspace/shortcuts.cljs
@@ -284,8 +284,8 @@
                           :subsections [:tools]
                           :fn #(emit-when-no-readonly (dw/toggle-lock-selected))}
 
-   :toggle-lock-size     {:tooltip (ds/meta (ds/alt "L"))
-                          :command (ds/c-mod "alt+l")
+   :toggle-lock-size     {:tooltip (ds/shift "L")
+                          :command "shift+l"
                           :subsections [:tools]
                           :fn #(emit-when-no-readonly (dw/toggle-proportion-lock))}
 
@@ -514,6 +514,10 @@
                           :subsections [:navigation-workspace]
                           :fn #(st/emit! (dw/select-next-shape))}
 
+   :select-parent-layer  {:tooltip (ds/shift ds/enter)
+                          :command "shift+enter"
+                          :subsections [:navigation-workspace]
+                          :fn #(emit-when-no-readonly (dw/select-parent-layer))}
    ;; SHAPE