+++ CHANGES.md
@@ -4,7 +4,7 @@
 
 ### :sparkles: New features
 
-- Navigate up in layer lierarchy with Shift+Enter shortcut [Taiga #5734](https://tree.taiga.io/project/penpot/us/5734)
+- Navigate up in layer hierarchy with Shift+Enter shortcut [Taiga #5734](https://tree.taiga.io/project/penpot/us/5734)
 
 ## 1.19.1